refactor: view 패키지 분리 및 CardDeck 인터페이스 도입
