+++ src/main/kotlin/blackjack/domain/CardDeck.kt
@@ -1,0 +1,6 @@
+package blackjack.domain
+
+fun interface CardDeck {
+
+    fun next(): Card
+}
