+++ src/main/kotlin/blackjack/domain/Player.kt
@@ -2,15 +2,15 @@ package blackjack.domain
 
 class Player(
     val name: String,
-    private val deck: Iterator<Card>,
+    private val cardDeck: CardDeck,
 ) {
-    private val cards = Cards(deck.next(), deck.next())
+    private val cards = Cards(cardDeck.next(), cardDeck.next())
     val hands
         get() = cards.values
 
     fun obtain() {
         require(isObtainable()) { "카드를 획득할 수 없습니다." }
-        cards.add(deck.next())
+        cards.add(cardDeck.next())
     }
 
     fun sumOfCards(): Int {
