+++ src/main/kotlin/blackjack/view/OutputView.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.view
 
 import blackjack.domain.Card
 import blackjack.domain.Number
